+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 ##### Features:
 - Rename [Index.php](src/ODM/Id/PrimaryKey.php) to [PrimaryKey.php](src/ODM/Id/PrimaryKey.php).
 
+### v0.1.2
+##### Features:
+- Added [FixturesTrait.php](src/ODM/Test/Helper/FixturesTrait.php).
+
+### v0.1.1
+##### Features:
+- Rename [Index.php](src/ODM/Id/PrimaryKey.php) to [PrimaryKey.php](src/ODM/Id/PrimaryKey.php).
+
 ### v0.1.0
 ##### Features:
 - Base implementation.